make scripts independant
